% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_database.R
\name{calc_shrub_density}
\alias{calc_shrub_density}
\title{Calculate density of shrub}
\usage{
calc_shrub_density(width = 2, length = 4, height = 1)
}
\arguments{
\item{width}{Width of shrub (numeric, default 2)}

\item{length}{Length of shrub (numeric, default 4)}

\item{height}{height of shrub (numeric, default 1)}
}
\value{
density of shrub (numeric, default 1)
}
\description{
Calculate density of shrub
}
