% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_nmds.R
\name{nmdsplot}
\alias{nmdsplot}
\title{This function performs nonmetric multidimensional scaling (NMDS) on a community data.frame and plots the NMDS to visualize dissimilarities.}
\usage{
nmdsplot(x, distance = "bray", autotransform = FALSE)
}
\arguments{
\item{x}{species community data.frame with species as cols and sites as rows}

\item{distance}{dissimilarity index used}

\item{autotransform}{automatic transformations of data points to whatever is most appropriate. To analyze raw data, set to =FALSE.}
}
\value{
nmds plot of data.frame
}
\description{
This function performs nonmetric multidimensional scaling (NMDS) on a community data.frame and plots the NMDS to visualize dissimilarities.
}
