% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Hillnumbers.R
\name{simpson}
\alias{simpson}
\title{This function incorporates Simpson diversity and reports the effective number of species as Hill Numbers of the order q.}
\usage{
simpson(x, q = 2, datatype = "abundance")
}
\arguments{
\item{x}{matrix, data.frame, vector, or list of species abundances or incidence frequencies to use}

\item{q}{numeric value specifying the diversity order of Hill number}

\item{datatype}{individual based abundance data}
}
\value{
effective number of species
}
\description{
This function incorporates Simpson diversity and reports the effective number of species as Hill Numbers of the order q.
}
